@@ -3,6 +3,8 @@
 var React = require('react');
 var assign = require('object-assign');
 var normalize = require('react-style-normalizer');
+var createClass = require("create-react-class");
+var DOM = require("../domFactories");
 
 var TEXT_ALIGN_2_JUSTIFY = {
     right: 'flex-end',
@@ -18,9 +20,9 @@ function copyProps(target, source, list) {
     });
 }
 
-var PropTypes = React.PropTypes;
+var PropTypes = require("prop-types");
 
-var Cell = React.createClass({
+var Cell = createClass({
 
     displayName: 'ReactDataGrid.Cell',
 
@@ -123,11 +125,20 @@ var Cell = React.createClass({
             }
         };
 
-        var content = props.renderCell ? props.renderCell(contentProps, text, props) : React.DOM.div(contentProps, text);
+        var content = props.renderCell ? props.renderCell(contentProps, text, props) : DOM.div(contentProps, text);
 
         var renderProps = assign({}, props);
 
+        delete renderProps.contentPadding;
         delete renderProps.data;
+        delete renderProps.defaultStyle;
+        delete renderProps.firstClassName;
+        delete renderProps.header;
+        delete renderProps.lastClassName;
+        delete renderProps.renderCell;
+        delete renderProps.renderText;
+        delete renderProps.rowIndex;
+        delete renderProps.textPadding;
 
         return React.createElement(
             'div',
