@@ -4,6 +4,7 @@ var React = require('react');
 var assign = require('object-assign');
 var ReactMenu = require('react-menus');
 var findDOMNode = require('react-dom').findDOMNode;
+var DOM = require("./domFactories");
 
 function stopPropagation(event) {
     event.stopPropagation();
@@ -19,7 +20,7 @@ module.exports = {
 
         var type = column.type || 'string';
 
-        return FILTER_FIELDS[type] || React.DOM.input;
+        return FILTER_FIELDS[type] || DOM.div;
     },
 
     getFilterField: function getFilterField(props) {
