@@ -13,6 +13,8 @@ var findIndexByName = require('../utils/findIndexByName');
 var Cell = require('../Cell');
 var setupColumnDrag = require('./setupColumnDrag');
 var setupColumnResize = require('./setupColumnResize');
+var createClass = require("create-react-class");
+var propTypes = require("prop-types");
 
 var normalize = require('react-style-normalizer');
 
@@ -56,12 +58,12 @@ function getDropState() {
     };
 }
 
-module.exports = React.createClass({
+module.exports = createClass({
 
     displayName: 'ReactDataGrid.Header',
 
     propTypes: {
-        columns: React.PropTypes.array
+        columns: propTypes.array
     },
 
     onDrop: function onDrop(event) {
@@ -329,7 +331,7 @@ module.exports = React.createClass({
                 selected: visible ? React.createElement(
                     'span',
                     { style: { fontSize: '0.95em' } },
-                    '✓'
+                    '\u2713'
                 ) : '',
                 label: column.title,
                 fn: this.toggleColumn.bind(this, column)
