@@ -1,7 +1,7 @@
 'use strict';
 
 module.exports = function (props, state) {
-                    var selected = props.selected == null ? state.defaultSelected : props.selected;
+    var selected = props.selected == null ? state.defaultSelected : props.selected;
 
-                    return selected;
+    return selected;
 };
