@@ -6,16 +6,14 @@ var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
-var _reactLoadMask = require('react-load-mask');
-
-var _reactLoadMask2 = _interopRequireDefault(_reactLoadMask);
-
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 require('es6-promise').polyfill();
 
 var assign = require('object-assign');
-
+var LoadMask = function LoadMask() {
+    return null;
+};
 var Region = require('region');
 
 var PaginationToolbar = _react2.default.createFactory(require('./PaginationToolbar'));
@@ -37,6 +35,9 @@ var getGroupedRows = require('./render/getGroupedRows');
 var renderMenu = require('./render/renderMenu');
 
 var preventDefault = require('./utils/preventDefault');
+var createClass = require("create-react-class");
+var propTypes = require("prop-types");
+var DOM = require("./domFactories");
 
 var isArray = Array.isArray;
 
@@ -83,32 +84,32 @@ function findColumn(columns, column) {
     }
 }
 
-module.exports = _react2.default.createClass({
+module.exports = createClass({
 
     displayName: 'ReactDataGrid',
 
     mixins: [require('./RowSelect'), require('./ColumnFilter')],
 
     propTypes: {
-        loading: _react2.default.PropTypes.bool,
-        virtualRendering: _react2.default.PropTypes.bool,
+        loading: propTypes.bool,
+        virtualRendering: propTypes.bool,
 
         //specify false if you don't want any column to be resizable
-        resizableColumns: _react2.default.PropTypes.bool,
-        filterable: _react2.default.PropTypes.bool,
+        resizableColumns: propTypes.bool,
+        filterable: propTypes.bool,
 
         //specify false if you don't want column menus to be displayed
-        withColumnMenu: _react2.default.PropTypes.bool,
-        cellEllipsis: _react2.default.PropTypes.bool,
-        sortable: _react2.default.PropTypes.bool,
-        loadMaskOverHeader: _react2.default.PropTypes.bool,
-        idProperty: _react2.default.PropTypes.string.isRequired,
+        withColumnMenu: propTypes.bool,
+        cellEllipsis: propTypes.bool,
+        sortable: propTypes.bool,
+        loadMaskOverHeader: propTypes.bool,
+        idProperty: propTypes.string.isRequired,
 
         //you can customize the column menu by specifying a factory
-        columnMenuFactory: _react2.default.PropTypes.func,
-        onDataSourceResponse: _react2.default.PropTypes.func,
-        onDataSourceSuccess: _react2.default.PropTypes.func,
-        onDataSourceError: _react2.default.PropTypes.func,
+        columnMenuFactory: propTypes.func,
+        onDataSourceResponse: propTypes.func,
+        onDataSourceSuccess: propTypes.func,
+        onDataSourceError: propTypes.func,
 
         /**
          * @cfg {Number/String} columnMinWidth=50
@@ -375,7 +376,7 @@ module.exports = _react2.default.createClass({
     },
 
     prepareFooter: function prepareFooter(props, state) {
-        return (props.footerFactory || _react2.default.DOM.div)({
+        return (props.footerFactory || DOM.div)({
             className: 'z-footer-wrapper'
         });
     },
@@ -422,7 +423,7 @@ module.exports = _react2.default.createClass({
         var loadMask;
 
         if (props.loadMaskOverHeader) {
-            loadMask = _react2.default.createElement(_reactLoadMask2.default, { visible: props.loading });
+            loadMask = _react2.default.createElement(LoadMask, { visible: props.loading });
         }
 
         var paginationToolbar;
