@@ -5,7 +5,9 @@ var renderMenu = require('./renderMenu');
 var renderRow = require('./renderRow');
 var tableStyle = require('./tableStyle');
 var slice = require('./slice');
-var LoadMask = require('react-load-mask');
+var LoadMask = function LoadMask() {
+    return null;
+};
 
 function getData(props) {
 
