@@ -7,6 +7,8 @@ var assign    = require('object-assign')
 var Toolbar   = require('react-simple-toolbar')
 var Region    = Toolbar.Region
 var normalize = require('react-style-normalizer')
+var createClass = require("create-react-class");
+var DOM = require("./domFactories");
 
 var WHITESPACE = '\u00a0'
 function sortAsc(a, b){
@@ -112,7 +114,7 @@ var defaultStyles = {
 	// gotoNext: { marginLeft: 10}
 }
 
-module.exports = React.createClass({
+module.exports = createClass({
 
 	displayName: 'PaginationToolbar',
 
@@ -294,7 +296,7 @@ module.exports = React.createClass({
 			onChange: this.handleInputChange
 		}, otherProps)
 
-		var defaultFactory = React.DOM.input
+		var defaultFactory = DOM.input
 		var factory = props.pageInputFactory || defaultFactory
 
 		var result = factory(inputProps)
@@ -319,7 +321,7 @@ module.exports = React.createClass({
 			children: options
 		}
 
-		var defaultFactory = React.DOM.select
+		var defaultFactory = DOM.select
 		var factory = props.pageSizeSelectFactory || defaultFactory
 
 		var result = factory(selectProps)
