@@ -2,13 +2,15 @@
 
 var React  = require('react')
 var assign = require('object-assign')
+var createClass = require("create-react-class");
+var propTypes = require("prop-types");
 
-module.exports = React.createClass({
+module.exports = createClass({
 
     displayName: 'ReactDataGrid.ResizeProxy',
 
     propTypes: {
-        active: React.PropTypes.bool
+        active: propTypes.bool
     },
 
     getInitialState: function(){
