@@ -5,7 +5,7 @@ var renderMenu = require('./renderMenu')
 var renderRow  = require('./renderRow')
 var tableStyle  = require('./tableStyle')
 var slice  = require('./slice')
-var LoadMask = require('react-load-mask')
+var LoadMask = () => null;
 
 function getData(props){
 
